Apply useSyncExternalStore for useGetSweetyState and useWatchSweety
