+++ package.json
@@ -53,14 +53,16 @@
     "@babel/preset-react": "^7.16.0",
     "@babel/preset-typescript": "^7.16.0",
     "@rollup/plugin-node-resolve": "^13.0.6",
-    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.0.0",
     "@testing-library/react-hooks": "^7.0.2",
     "@types/jest": "^27.4.0",
-    "@types/react": "^17.0.43",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "@types/use-sync-external-store": "^0.0.3",
     "@typescript-eslint/eslint-plugin": "^5.4.0",
     "@typescript-eslint/parser": "^5.4.0",
+    "@vitejs/plugin-react": "^1.3.0",
     "eslint": "^8.2.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.25.3",
@@ -72,12 +74,14 @@
     "jest": "^27.4.7",
     "prettier": "^2.4.1",
     "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-test-renderer": "^18.0.0",
     "rollup": "^2.60.0",
     "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-replace": "^2.2.0",
     "rollup-plugin-terser": "^7.0.2",
     "rollup-plugin-typescript2": "^0.31.0",
-    "typescript": "^4.5.2"
+    "typescript": "^4.5.2",
+    "vite": "^2.9.1"
   }
 }
