+++ src/useWatchSweety.ts
@@ -1,7 +1,10 @@
-import { useRef, useReducer, useEffect } from "react"
+import { useRef, useEffect, useCallback } from "react"
+import { useSyncExternalStoreWithSelector } from "use-sync-external-store/with-selector"
 
-import { Compare, modInc, isEqual } from "./utils"
+import { Compare, isEqual } from "./utils"
 import { WatchContext } from "./WatchContext"
+
+const identity = <T>(value: T): T => value
 
 /**
  * A hook that subscribes to all `Sweety#getState` execution involved in the `watcher` call.
@@ -19,45 +22,68 @@ export function useWatchSweety<T>(
   watcher: () => T,
   compare?: null | Compare<T>,
 ): T {
-  const [, render] = useReducer(modInc, 0)
-
-  const valueRef = useRef<T>()
-  const watcherRef = useRef<() => T>()
-  // workaround to handle changes of the watcher returning value
-  if (watcherRef.current !== watcher) {
-    valueRef.current = WatchContext.executeWatcher(watcher)
-  }
-
-  // permanent ref
   const contextRef = useRef<WatchContext>()
-  if (contextRef.current == null) {
-    contextRef.current = new WatchContext(() => {
-      const currentValue = valueRef.current!
-      const nextValue = WatchContext.executeWatcher(watcherRef.current!)
+  // const watcherRef = useRef(watcher)
 
-      if (!compareRef.current(currentValue, nextValue)) {
-        valueRef.current = nextValue
-        render()
-      }
-    })
+  if (contextRef.current == null) {
+    contextRef.current = new WatchContext()
   }
 
-  const compareRef = useRef(compare ?? isEqual)
-  useEffect(() => {
-    compareRef.current = compare ?? isEqual
-  }, [compare])
+  const getState = useCallback(() => {
+    return WatchContext.executeWatcher(watcher)
+  }, [watcher])
 
   useEffect(() => {
-    watcherRef.current = watcher
     contextRef.current!.activate(watcher)
   }, [watcher])
 
-  // cleanup everything when unmounts
-  useEffect(() => {
-    return () => {
-      contextRef.current!.cleanup()
-    }
-  }, [])
+  return useSyncExternalStoreWithSelector(
+    useCallback((fire) => contextRef.current!.subscribe(fire), []),
+    getState,
+    getState,
+    identity,
+    compare ?? isEqual,
+  )
 
-  return valueRef.current!
+  // const [, render] = useReducer(modInc, 0)
+
+  // const valueRef = useRef<T>()
+  // // const watcherRef = useRef<() => T>()
+  // // workaround to handle changes of the watcher returning value
+  // if (watcherRef.current !== watcher) {
+  //   valueRef.current = WatchContext.executeWatcher(watcher)
+  // }
+
+  // // permanent ref
+  // // const contextRef = useRef<WatchContext>()
+  // if (contextRef.current == null) {
+  //   contextRef.current = new WatchContext(() => {
+  //     const currentValue = valueRef.current!
+  //     const nextValue = WatchContext.executeWatcher(watcherRef.current)
+
+  //     if (!compareRef.current(currentValue, nextValue)) {
+  //       valueRef.current = nextValue
+  //       render()
+  //     }
+  //   })
+  // }
+
+  // const compareRef = useRef(compare ?? isEqual)
+  // useEffect(() => {
+  //   compareRef.current = compare ?? isEqual
+  // }, [compare])
+
+  // useEffect(() => {
+  //   watcherRef.current = watcher
+  //   contextRef.current.activate(watcher)
+  // }, [watcher])
+
+  // // cleanup everything when unmounts
+  // useEffect(() => {
+  //   return () => {
+  //     contextRef.current.cleanup()
+  //   }
+  // }, [])
+
+  // return valueRef.current!
 }
