+++ src/utils.ts
@@ -55,41 +55,3 @@ export const overrideCompare = <T>(
 export const noop: VoidFunction = () => {
   // do nothing
 }
-
-/**
- * @private
- */
-export const identity = <T>(value: T): T => value
-
-/**
- * @private
- */
-export const warning = (message: string): void => {
-  /* istanbul ignore next */
-  if (
-    (process.env.NODE_ENV === "development" ||
-      process.env.NODE_ENV === "test") &&
-    typeof console !== "undefined" &&
-    // eslint-disable-next-line no-console
-    typeof console.error === "function"
-  ) {
-    // eslint-disable-next-line no-console
-    console.error(message)
-  }
-  /* eslint-enable no-console */
-  try {
-    // This error was thrown as a convenience so that if you enable
-    // "break on all exceptions" in your console,
-    // it would pause the execution at this line.
-    throw new Error(message)
-  } catch {
-    // do nothing
-  }
-}
-
-/**
- * @private
- */
-export const modInc = (x: number): number => {
-  return (x + 1) % 123456789
-}
